+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## unreleased
 
+- fix: Use RateLimitCategoryError for events #470
 - feat: Store SentryEnvelopes in extra path #468
+- feat: Adds setUser to SentrySDK and SentryHub #467
 - feat: Add auto session starting for macOS #463
 - fix: Take release name from options #462